phil modifications to kathik refactor
